Update patch set 1
Patch Set 1:

I was listening to the recorded session that referenced this gerri patchset. It appears to me that this "dot" file defines "what is expected". Is that accurate? If so, how do we compare "expected" to actual or even capture actual as code is developed and committed?

Patch-set: 1
Reviewer: Gerrit User 1675 <1675@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0