Update patch set 1
Patch Set 1:

If we want to ensure all dependencies are documented, don't we need to formalize how dependencies are specified (i.e. in a file in each project as opposed to a wiki?) and then a security mechanism so that only defined dependencies can be invoked, so that one module cannot call another until a dependencies is officially defined?

Patch-set: 1