Update patch set 3
Patch Set 3:

Hi Victor - that makes sense, though I think if the unit_testing file is intende to be called only by Vagrant maybe we should put it in a folder called vagrant_utils or something. It's a little confusing having it in the main directory, it seems like it should be called by the user to perform unit tests.

I'll update this patch to remote the unneeded instructions rather than adding options.

Patch-set: 3
Label: Code-Review=0