Update patch set 4
Patch Set 4:

The thing that I've tried to achieve here is that only if you want to clone you'll need to compile and vice versa. In my case we don't want the code to be on our vagrant.
I can't seem to understand what is the problem of giving the user the ability to choose if he wants to clone and compile the code or not?

Patch-set: 4
Reviewer: Gerrit User 1341 <1341@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0