Update patch set 4
Patch Set 4:

the question is is their a point in separating the two variables compile and clone?
in addition i wanted to know are you aware if any one except us is using this vagrant?
and how are they working with it are they cloning into it?
and i belive that the clone flag will serve all components and allow them more flexibility to use this as they see fit.

Patch-set: 4
Reviewer: Gerrit User 69 <69@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0