Update patch set 14
Patch Set 14:

Well that got a little closer. At least we saw a 404 error eventually in the first health check call. Jorge, how long should we wait for the drools pdp to be ready? Are the logs anywhere that we can dump to see what it is reporting?

Patch-set: 14
Label: Code-Review=0