Update patch set 2
Patch Set 2:

Not sure what exactly is wrong but Gerrit will not allow me to merge this due to "missing dependencies".  If I try to rebase via the UI then it complains of merge conflicts.  You may need to re-submit this change on top of the latest master branch.

Patch-set: 2