Update patch set 2
Patch Set 2:

> two qustions :
 > 1) oom needs 40 g ram????

I tested with that configuration but I'm not completely sure if it's working properly, maybe that's the minimum memory required to have services up and running without any workload

 > 2) how can this be used as a local env for devalopment?

One alternative could be deploying OOM using the OpenStack provider[1] and the desired component locally. Another alternative is trying to use the all-in-one configuration and using the build functions to populate the local registry besides the change in OOM to use that local registry instead of external.

[1] http://onap.readthedocs.io/en/latest/submodules/integration.git/bootstrap/vagrant-onap/doc/source/features/openstack.html

Patch-set: 2