Update patch set 1
Patch Set 1:

Hey Victor, I think this is a good change. Since we're now including the all-in-one inventory in-tree here and it's used to deploy OpenStack services, do you think we should also add to the "enable" section to the globals file and allow the user to choose which services to enable and disable, and mention this option in the docs? This could be done in a follow-up patch, I wouldn't mind adding that.

Patch-set: 1