Optimization Policy CSIT tests
This CSIT tests requires an entry in the database.  I modified the
sql script in the PE packages.  That commit should be merged in order
for this to pass

I had the wrong variable name for ${policy_name} in the template.  I
fixed this but I am not able to test locally so I cant be 100% sure this
will resolve the issue

I added fixes for the 2 failures.  I think the createPolicy failed due
to an invalid model but I see the model being inserted in the pap log.
The only thing I can think of is that it does not like the camel case
for hpaPolicy and the all caps for CSIT version so I just made these all
lower case in the request template

Change-Id: I59fe578c57ab927fe356071a5b4a28bd987fafd7
Issue-ID: POLICY-639
Signed-off-by: Michael Mokry <mm117s@att.com>