Update patch set 2
Patch Set 2:

The only thing that comes to my mind is that the file should be sorted. Not sure whether it is required to pass the verification, but it's worth trying. Please resubmit after running "cat java-manifest.csv | sort -o java-manifest.csv" locally. Hopefully that will fix the problem.

Patch-set: 2
Label: Verified=0