Update patch set 5
Patch Set 5:

Looks like the failures are now due to the following error...

Received response from policy PE300 - Data Issue: Invalid Template.  The template name, ClosedLoopControlName was not found in the dictionary.

so it looks like the template for the ops policies is missing.. this is essentially the same kind of error as invalid model for Config policies.  The HPA create policy passed so thats good :) but now the getCOnfig call for it failed.. I should be able to fix that but I am not sure if anything changed for the OPS policies that would cause them to fail now.

Patch-set: 5