Update patch set 7
Patch Set 7:

Pam, I know what the cause of the failure is.. seems the config file is being creating with an added .xml.json in it.. it should just be .json file but I will have to look at the Optimization createPolicy code to fix it.  Do we need a bug for this?  Other than that this CSIT tests looks good and should pass once I fix the issue with creating the config file for OOF policy.

Patch-set: 7