Update patch set 3
Patch Set 3:

The test machine of Jenkins uses an older version of Docker (1.12) so it seems unhappy with the old syntax of `docker inspect`, NBI_IP will then be set to the full output of the `docker inspect` command instead of the container ip address. However no error has been raised because the for-loop at the end of the script forces the machine to wait *465 seconds* in total before launching the robot tests, more than enough for the nbi container to be ready for connections. This patch will fix that error and reduce the test duration.

Patch-set: 3
Reviewer: Gerrit User 3787 <3787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0