Update patch set 4
Patch Set 4:

it looks like the job is failing because they are not releasing images. how ever they are not supposed to be since we are working with the snapshots until the end of the release.

Patch-set: 4
Reviewer: Gerrit User 69 <69@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0