Update patch set 4
Patch Set 4:

That means that the image exists but is not a released image.  If you are ready to release the images, the PTL needs to send an email to LF helpdesk to request their release.

If you are not ready to release the images yet, don't update docker-manifest.csv yet.

Patch-set: 4