Update patch set 1
Patch Set 1:

Following up with LF helpdesk on the error, seems 2 images have not been properly released.

In the meamtime, let me ask you, the version-manifest/src/main/resources/docker-manifest-staging.csv, is it expected to be modified too?    Not sure if we did this one in the past.   I made the changes, but let me know otherwise.

Patch-set: 1