Update patch set 1
Patch Set 1:

This job is failing the verify, but it doesn't seem related to the changes I made.  It appears to be failing on the check-docker-images-released step and reporting that several images could not be found.  Can someone more familiar with this project take a look and let me know if there's something I'm missing?

Patch-set: 1
Reviewer: Gerrit User 3999 <3999@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0