Update patch set 1
Patch Set 1:

The staging CSV is still necessary so that integration can override certain images with the staging versions as necessary.  We don't expect a fix for the "one image multiple versions" issue with the CSV.  Ultimately the CSV needs to be converted to just a plain helm override file to deal with that issue.

If your team is currently working a bug fix and expect to release another version, then putting the staging version for that next release here will allow that staging version to be picked up by integration's staging deployments.  If you're not expecting to release any additional versions, then you should put the current release version in this CSV.

Patch-set: 1