Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Merging this was little hasty, we should have had the CI job merged first, then verify this and only then merge...
> 
> Is it possible to validate the CI job if the target source code isn't already merged?

For me the better approach is to merge the CI job first considering it valid once it passes LF verify job for the job validation itself, then issue 'recheck' here to validate the code that was uploaded to integration repo.

So in the end it's just the matter of what is more important to test first: CI job itself, or the code here. Merging it without verification was advocating for the first case, merging CI job first for the latter.

After CI job is merged just issue 'recheck' here.

Patch-set: 1