Update patch set 2
Patch Set 2: Verified+1

> Patch Set 2:
> 
> @bartek I think it would make sens to extend the json + python verification to this simulator

As for python linting it's a work in progress, please see my comment: https://jira.onap.org/browse/INT-1124?focusedCommentId=52530&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-52530

All json files in repo are in scope of the integration-master-json-lint job; it's just it was merged only yesterday (CI job in ci-management) after this review was uploaded. Rechecked however it's failing, I'm looking into it - it looks like it's the issue with LF tools itself because it fails in post stage:

---> python-tools-install.sh
Generating Requirements File
ERROR: Package 'zipp' requires a different Python: 3.5.2 not in '>=3.6'

It's not related to job config.

The JSON verification itself passed ok so marking this verified:

___________________________________ summary ____________________________________
  json: commands succeeded
  congratulations :)

Patch-set: 2