Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> e.g. "diff -u"
> 
> Is this shell script referenced somewhere or it's just meant to be run manually? Can't it leverage version control or at least patching instead of such extensive and unpredictable (from resulting output perspective) use of stream editor? Also looking at line 26 where it bounds the flow to a specific commit it seems hard to maintain in future releases.

Hi,

This shell script is just to be run manually.

It bounds to a specific commit is to keep the patch files are always available.

If doesn't bound to a specific commit, the patch files may be invalid when the source CBA files in CDS code are updated in the future.

I will use patching instead of sed command in the next patch set.

Patch-set: 6
Label: Code-Review=0