Update patch set 1
Patch Set 1: Code-Review-1

This port is not for testing purpose only.
As per discussion during yesterday SECCOM our understanding is that the reason why this port is required is because xNF simulators that are used for testing are not capable of sending events using HTTPs interface.

Unfortunately this means that our simulators xNFs are incompatible with ONAP VNF requirements which is a terrible practice. That's why we believe that this port should not get a waiver unless there is some additional rationale provided

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1