Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This port is not for testing purpose only.
> As per discussion during yesterday SECCOM our understanding is that the reason why this port is required is because xNF simulators that are used for testing are not capable of sending events using HTTPs interface.
> 
> Unfortunately this means that our simulators xNFs are incompatible with ONAP VNF requirements which is a terrible practice. That's why we believe that this port should not get a waiver unless there is some additional rationale provided

PNF simulator will be fixed for Frankfurt (WIP)
for vFW and vCPE, which are using the VES for the closed loops, nothing has been initiated yet.
Based on PNF feedback it may be relatively quick but I do not have the visibility.

I suggest we close these two ports (and then abandon this patch  - no waiver)
Integration document that for Frankfurt, for testing only we need to open these ports.
We reopen the JIRA (I closed them because I had no resources and believed it was up to the use case team to update their simulators..but use case and integration teams are often the same people at the end.

https://jira.onap.org/browse/INT-1268 (Epic)
https://jira.onap.org/browse/INT-1317 (the only one open and WIP)
https://jira.onap.org/browse/INT-1319
https://jira.onap.org/browse/INT-1318

The tests were not in the CI chains, so I do not think it is a problem to add a manual step in the documentation.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0