Update patch set 1
Abandoned

See discussions in the comments. Use case have to update their simulators to comply with security requirements. JIRAs have been reopened for Guilin. Documentation shall be updated to reflect that the port must be manually opened to perform the test for the use cases that did not implement the migration to HTTPS/VES yet

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon