Update patch set 2
Patch Set 2:

My git folder got messed up, so I lost this contribution on my local machine. If you can merge it as is, I will rework it, otherwise I lose everything. The rst file should be fine, but it's pointing to a wrong link, which causes the issue. It should be an easy fix.

Patch-set: 2