Update patch set 4
Patch Set 4:

(11 comments)

Corrected most of issues.
As far as I'm aware the only leftover parts  to finish docs are:
  * FIXMES - all relate to updating logs/server outputs sections against RC2 Frankfurt
  * UAT - not part of this review

If you still have some issues - please point them.
If not, I guess we can have this commit merged `eventually` if we were in hurry. If we are not, FIXMES could be also covered under this review (need RC2 ready).

Please say what is point of view over it.

Patch-set: 4