Update patch set 2
Patch Set 2:

I am still not clear on the source of the need, I found the JIRA but did not find any requirement
it has never been discussed in my project, so it is a request from LF to save resources?

if so I think that the history is not the main problem regarding the size of the repositories.
if it would be easy to create repositories, we would have more small repositories and not few monster ones used as shared drive that could be long to clone

In your patch, you looked for all the instances, in the use case documentation (rst) It will not save anything.
on  bootstrap and deployment, why not, but today the full deployments are not done from here in CI so it will not save time (and money) on Vexxhost here AFAI

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>