Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> would it make sense for G to create a dedicated repository integration/simulators/mass-pnf-simulator as we have dc-simulator or pnf-simulator? It could make the simu more visible, no?

In Integration repo there are several alike projects:

datafilecollector-testharness
emssimulator
hvvessimulator
mass-pnf-sim
netconf-pnp-simulator
pmsh-pnf-sim
pnf-onboarding
prov-mns-provider
sniroemulator

Do we want to move all of them to a separate repositories? If not then mass PNF simulator should be no exception in this regard i think. And if a decision on moving a particular subproject to a separate repo is to be made just on a factor of it's current support status then not sure either as now I'm contributing a lot but in few months it may stall again. 

So personally I'd leave it in Integration for now with a possibility to move it to a separate repo if some other factors emerge.

Patch-set: 1
Reviewer: Gerrit User 5097 <5097@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>