Update patch set 2
Patch Set 2: Code-Review+2

> Patch Set 2:
> 
> This commit is only adding more Python native way of defining the runtime environment for the script - previously it was required to install packages (kubernetes and colorama) currently by hand, now tox takes care of it by creating virtualenv.
> 
> Hence I'm not sure i understand the context of "final status" "PASS or FAIL"?

agree and inline with you, it is better to do that here (rather at the xtesting level)
As it was not done here I added the requirements in the xtesting docker file including this test
https://git.onap.org/integration/xtesting/tree/infra-healthcheck/requirements.txt
xtesting is ~ virtual env (it is a docker) but it is better when python packages are properly packaged earlier

the question regarding the success criteria is more dealing wit the test itself
if you run this script, you got logs
but at the end you do not know if the test is PASS or FAIL, you have to compare the number of ingress and number of port I think
to include it in CI, we should get a clear statement that could be a log "Test PASS" or the return code of the test (return 0 if PASS)
but as far as I remember we do not have such things, so we can execute the test and see the logs. The success criteria has not been set.

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2