Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (4 comments)
> 
> I recall letting you know that I put my checks in test/security but I'm not sure if it's the best place for license compliance utilities. I don't know which Xtesting category it falls into (https://git.onap.org/integration/xtesting/tree/), either - let's hear if anyone has strong preference.
> 
> -1 score is only for the lack of usage docs outside commit history

for the location...
all the tests we run in our CI are in xtesting dockers
but here we would like to run the test in jenkins LF chain..so it is not necessarily mandatory to onboard the test in a xtesting.

It could be in security or we could create a new one (sanity?legal?) - remember xtesting is "just" a python lib allowing to harmonize the way we are running the tests in order to simplify the integration in any CI..

I assume the script could be run directly from the jenkins jjb and hosted in ci-management
but for the PoC it is fine to host it in integration
Moreover it deals with only 2 files so we can easily wget them if needed.

Patch-set: 1