Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> I think you missed a carriage return - the title is a bit long.. 😊
> would it be possible to plan a demo of these tests during an integration weekly meeting?
> not this week as it is public holiday in Poland but maybe next week
> 
> WOuld some ofg the TC be suitable for End to End tests in CI/CD chains (in addition of functional/CSIT tests)

Hi Morgan, 
I've updated commit message. Hope this will be suitable. 
It would be pleasure to prepare a demo of these tests during integration weekly meeting.
Next week will be fine.
For now I'm not sure if the prepared TC will be suitable for End to End tests. There are
6 TC prepared to cover CSIT tests with HTTP. Corresponding auto-tests are in the autotests
directory, but for now I'm not sure how or if they are used in CI/CD chains.

Patch-set: 2