Update patch set 7
Patch Set 7:

(7 comments)

thanks for the update
as said in the review, there are still lots of ref to the wiki that is not really under control, it would be better to import doc from wiki to put it in official doc, as this doc is fully versioned and cannot be amended without a patch.

otherwise minor comments on the format and the choice to remove the hyperlinks in the text?
the current render (with this patch)

https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/rtdv3-global-verify-master/3778/docs/docs_E2E_network_slicing.html#docs-e2e-network-slicing

versus the old official version
https://docs.onap.org/projects/onap-integration/en/latest/docs_E2E_network_slicing.html#docs-e2e-network-slicing

I think hyperlinks look better for an official documentation

Patch-set: 7
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>