Update patch set 1
Patch Set 1:

Was this tool moved to Orange's gitlab account space? If so is this a really good idea? Since this tool belongs to Onap ecosystem shouldn't it stay within Onap's Gerrit? Not this repo necessarily ofc. If it's home is Gitlab it will be harder for Onap contributors to eventually contribute to it, no?

Patch-set: 1
Reviewer: Gerrit User 5097 <5097@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>