[CODESEARCH] Rework how --git option works
"--git" will not only affect the "base-url" in Hound config for
code browser but also the "url" for the repository clone string.
This is the most sane behavior as git protocol is the most efficient
one in terms of speed.

"git" and "ssh" options are also set to be mutually exclusie as
according to the above change using git protocol based clone and
passing ssh credentials would not make sense as git protocol does
not support any kind of authentication.

Change-Id: I9897bc2aec2707ca4ed7543b1576686fb8e2b2da
Issue-ID: INT-1940
Signed-off-by: Bartek Grzybowski <b.grzybowski@partner.samsung.com>