[CODESEARCH] Rework the nameserver provisioning
Default OS resolver is set using native systemd-resolved
facilities in a way that it will be persisted after box
reload.

Default nameserver is set within Vagrantfile and can be altered
if needed.

Vagrant-libvirt specific resolver settings were removed in favor
of a general one. As of now it should be possible to also run this
box using Virtualbox provider.

Change-Id: Icdd1eb8bf3677c4616b1ac7d6b3387e0f4c7535a
Issue-ID: INT-1940
Signed-off-by: Bartek Grzybowski <b.grzybowski@partner.samsung.com>