Update patch set 1
Patch Set 1:

this ansible roles to deploy rke look similar to the ansible collection https://gitlab.com/Orange-OpenSource/lfn/infra/rke2_automatic_installation_collection

I understand that historically such ansible roles/collections were hosted in the integration repo.
It would probably make sense to create a dedicated repository and converge with the Open Source repo on gitlab (which is not directly under the LFN authority for several historical reasons even if we use it for ours deployments/CI).
For istanbul I assume we keep everything as it is, but it would be interesting to have a discussion OOM/Integration to see how we could improve that in the future and have 1 ansible collection

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>