Update patch set 1
Patch Set 1:

(6 comments)

thanks for the update
minor suggestion + question on the tests

the verify - is due to a broken link that is not really broken..but I assume it is too long to answer. Anyway it has nothing to deal with the patch.
I will remove this reference, then rebasing this patch shall lead to a +1

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>