Adds 2 upper-constraints to pin all the dependencies.
It modifies a little bit the official upper-constraints.txt from
OpenStack due to Sphinx (4.2.0 as asked by lfdocs-conf vs 3.5.2) and
due to setuptools (58.1.0 as used in ONAP gates vs 54.1.1).

In a long run, docs/upper-constraints.os.txt should be removed once ONAP
is synced with OpenStack. For it's part, docs/upper-constraints.onap.txt
should be centralized for all ONAP projects and then removed from this
particular project.

It should be noted that lfdocs-conf still mismatches concrete and
abstract deps.

Issue-ID: DOC-765

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
Change-Id: Iae808297484f4798de82a43597ccad7905ff4c94