Update patch set 1
Patch Set 1: Code-Review+2

I know that it is not easy, but it would make sense to spend some time to refactor a little bit the doc - maybe by newcommer (integration?)

In fact the incremental documentation (what has been done in frankfurt, Guilin, Honolulu, Istanbul...) is sometimes hard to follow when you are not involved in the project. 
A high level description independent + a roadmap of the changes would be nice.

Moreover for a mature use case like this one, it would be nice to have automation in CI, it will secure the maintainability of the use case and simplify the reproducibility in any lab.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2